Backend/app/requirements.txt
beautifulsoup4==4.12.0
Flask==2.2.3
flask_cors==3.0.10
pydantic==1.10.7
pymongo==4.3.3
python-owasp-zap-v2.4
